requirements.txt
pillow
pywikibot
